[hw09-ts] Add TS to project - step 1
