+++ node-2023-11/hw04_les07-webServers/.eslintrc.js
@@ -1,0 +1,27 @@
+module.exports = {
+    env: {
+        browser: true,
+        es6: true
+    },
+    globals: {
+        Atomics: 'readonly',
+        SharedArrayBuffer: 'readonly'
+    },
+    parser: '@typescript-eslint/parser',
+    parserOptions: {
+        ecmaVersion: 2018,
+        sourceType: 'module'
+    },
+    plugins: ['@typescript-eslint', 'prettier'],
+    rules: {
+        semi: ['warn', 'always'],
+        'prettier/prettier': [
+            'error',
+            {
+                endOfLine: 'auto'
+            }
+        ],
+        '@typescript-eslint/explicit-function-return-type': 0,
+        'react/prop-types': 0
+    }
+};
