+++ node-2023-11/hw04_les07-webServers/.prettierignore
@@ -1,0 +1,5 @@
+.history
+.vscode
+node_modules
+public
+*.config.js
