+++ node-2023-11/hw04_les07-webServers/api-tests/constants.ts
@@ -1,0 +1,9 @@
+import db from '../services/db.service';
+
+export const PETER_ID = db.to24Str('01');
+export const NICK_ID = db.to24Str('02');
+export const DELME_ID = db.to24Str('03');
+export const TOM_ID = db.to24Str('04');
+export const MATH_ID = db.to24Str('0101');
+export const HISTORY_ID = db.to24Str('0102');
+export const FILE_ID = db.to24Str('0201');
