--- node-2023-11/hw04_les07-webServers/babel.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-    plugins: ['@babel/plugin-proposal-class-properties'],
-    env: {
-        production: {
-            plugins: []
-        },
-        development: {
-            plugins: []
-        }
-    }
-};
