+++ node-2023-11/hw04_les07-webServers/babel.config.json
@@ -1,0 +1,27 @@
+{
+    "presets": [
+      [
+        "@babel/env",
+        {
+          "targets": {
+            "chrome": "41",
+            "node": "12"
+          },
+          "debug": false
+        }
+      ],
+      "@babel/typescript"
+    ],
+    "plugins": [
+      [
+          "module-resolver",
+          {
+              "alias": {
+                  "@src": "./src",
+                  "@": "./"
+                }
+          }
+      ]
+  ]
+
+}
