+++ node-2023-11/hw04_les07-webServers/bin/www.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 /**
  * Module dependencies.
  */
