+++ node-2023-11/hw04_les07-webServers/controllers/adminFiles.controller.ts
@@ -1,4 +1,4 @@
-const File = require('../models/File').File;
+import { File } from '../models/File';
 
 function get(req, res, next) {
     File.find({})
