+++ node-2023-11/hw04_les07-webServers/package.json
@@ -7,6 +7,34 @@
     "watch": "nodemon",
     "test": "jest"
   },
+  "devDependencies": {
+    "@babel/core": "^7.22.1",
+    "@babel/preset-env": "^7.22.4",
+    "@babel/preset-typescript": "^7.21.5",
+    "@types/faker": "^5.5.9",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^18.8.5",
+    "@typescript-eslint/eslint-plugin": "^5.60.0",
+    "@typescript-eslint/parser": "^5.60.0",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-module-resolver": "^5.0.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "faker": "^5.5.3",
+    "jest": "^27.3.1",
+    "node-loader": "^2.0.0",
+    "nodemon": "^2.0.4",
+    "prettier": "^2.4.1",
+    "redux": "4.0.5",
+    "redux-actions": "2.6.5",
+    "supertest": "^6.3.3",
+    "ts-node": "^10.3.0",
+    "tsconfig-paths": "^4.0.0",
+    "typescript": "^4.4.4",
+    "webpack": "^5.85.0",
+    "webpack-cli": "^5.1.1"
+  },
   "dependencies": {
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
@@ -22,16 +50,8 @@
     "ws": "^8.16.0",
     "yup": "^1.3.3"
   },
-  "devDependencies": {
-    "@babel/eslint-parser": "^7.23.3",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "babel-core": "^6.26.3",
-    "eslint": "^7.17.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-prettier": "^3.3.0",
-    "jest": "^29.7.0",
-    "nodemon": "^3.0.2",
-    "prettier": "^2.2.1",
-    "supertest": "^6.3.3"
+  "browser": {
+    "fs": "false",
+    "child_process": "false"
   }
 }
