+++ node-2023-11/hw04_les07-webServers/prettier.config.js
@@ -4,4 +4,4 @@ module.exports = {
     trailingComma: 'none',
     tabWidth: 4,
     semi: true
-};
+};
