+++ node-2023-11/hw04_les07-webServers/routes/adminFiles.ts
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 const { Privileges } = require('../constants');
-const db = require('../services/db.service');
+import db from '../services/db.service';
 const adminFilesController = require('../controllers/adminFiles.controller');
 
 router.get('/', db.checkAuth, db.hasOneOfPriv([Privileges.filesAdmin]), adminFilesController.get);
