+++ node-2023-11/hw04_les07-webServers/routes/adminUsers.ts
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-const db = require('../services/db.service');
+import db from '../services/db.service';
 const { Privileges } = require('../constants');
 const adminUsersController = require('../controllers/adminUsers.controller');
