+++ node-2023-11/hw04_les07-webServers/services/FileStorageService.ts
@@ -1,29 +1,28 @@
-const fs = require('fs');
+import fs from 'fs';
 var path = require('path');
 
 class FileStorageService {
-    dir = null;
-    constructor(dir) {
+    dir: string = '';
+    constructor(dir: string) {
         this.dir = dir;
     }
 
     async clearStorageDir() {
-        let dirHandle = null;
+        let dirHandle: fs.Dir;
         try {
             dirHandle = await this.openDirectory(this.dir);
         } catch (e) {
             fs.promises.mkdir(this.dir, { recursive: true });
             return;
         }
-
         for await (const dirent of dirHandle) {
             const fullFName = path.join(this.dir, dirent.name);
 
             fs.promises.unlink(fullFName);
         }
-    }
+}
 
-    openDirectory = async (path) => await fs.promises.opendir(path);
+    openDirectory = async (path: string) => await fs.promises.opendir(path);
 }
 
 module.exports = {
